+++ packages/realm_dart/test/realm_map_test.realm.dart
@@ -39,11 +39,14 @@ class Car extends _Car with RealmEntity, RealmObjectBase, RealmObject {
   static SchemaObject? _schema;
   static SchemaObject _initSchema() {
     RealmObjectBase.registerFactory(Car._);
-    return const SchemaObject(ObjectType.realmObject, Car, 'Car', [
+    return SchemaObject(ObjectType.realmObject, Car, 'Car', [
       SchemaProperty('make', RealmPropertyType.string, primaryKey: true),
       SchemaProperty('color', RealmPropertyType.string, optional: true),
     ]);
   }
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class EmbeddedValue extends _EmbeddedValue
@@ -72,11 +75,14 @@ class EmbeddedValue extends _EmbeddedValue
   static SchemaObject? _schema;
   static SchemaObject _initSchema() {
     RealmObjectBase.registerFactory(EmbeddedValue._);
-    return const SchemaObject(
+    return SchemaObject(
         ObjectType.embeddedObject, EmbeddedValue, 'EmbeddedValue', [
       SchemaProperty('intValue', RealmPropertyType.int),
     ]);
   }
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }
 
 class TestRealmMaps extends _TestRealmMaps
@@ -321,7 +327,7 @@ class TestRealmMaps extends _TestRealmMaps
   static SchemaObject? _schema;
   static SchemaObject _initSchema() {
     RealmObjectBase.registerFactory(TestRealmMaps._);
-    return const SchemaObject(
+    return SchemaObject(
         ObjectType.realmObject, TestRealmMaps, 'TestRealmMaps', [
       SchemaProperty('key', RealmPropertyType.int, primaryKey: true),
       SchemaProperty('boolMap', RealmPropertyType.bool,
@@ -372,4 +378,7 @@ class TestRealmMaps extends _TestRealmMaps
           optional: true, collectionType: RealmCollectionType.map),
     ]);
   }
+
+  @override
+  SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;
 }