+++ packages/realm_generator/lib/src/realm_model_info.dart
@@ -124,7 +124,7 @@ class RealmModelInfo {
       yield 'static SchemaObject _initSchema() {';
       {
         yield 'RealmObjectBase.registerFactory($name._);';
-        yield "return const SchemaObject(ObjectType.${baseType.name}, $name, '$realmName', [";
+        yield "return SchemaObject(ObjectType.${baseType.name}, $name, '$realmName', [";
         {
           yield* fields.map((f) {
             final namedArgs = {
@@ -146,6 +146,9 @@ class RealmModelInfo {
         yield ']);';
       }
       yield '}';
+      yield '';
+      yield '@override';
+      yield 'SchemaObject get objectSchema => RealmObjectBase.getSchema(this) ?? schema;';
     }
     yield '}';
   }